feat: app redesign (#78)
**Description**
In this pull request, we have undertaken a comprehensive rebranding of the application interface to reflect the updated brand identity. This includes updates to color schemes, typography and iconography to ensure a cohesive visual experience that aligns with the new branding guidelines. There have been no major changes to the overall layout of the application.

Closes #65 

**Other changes**
- None

**Tested**
Changes have been tested across various screen sizes and devices to ensure a consistent user experience. This includes manual testing of all key user flows.

**Please note:** While all main functionality has been tested and is working as expected, due to other changes that were made at the same time and have been merged to main, there may be some minor issues related to swapping(especially with Valora) that were not caught in other code review. These should be addressed separately.