chore: add pr template with checklist (#82)
### Description

This PR introduces a standardized PR template for all future pull requests in this repository.

### Other changes

No other changes are included in this PR; it's solely focused on adding the PR template.

### Tested

This PR is related to process and documentation and does not include code changes. Hence, it doesn't require the usual code testing. 

### Related issues

There is no related issue for this.

### Checklist before requesting a review

- [x] I have performed a self-review of my own code
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] The PR title follows the [conventions](https://www.notion.so/Git-Branching-and-Commit-Message-Conventions-18f66f7d06444cfcbac5725ffbc7c04a?pvs=4#9355048863c549ef92fe210a8a1298aa)
- [x] I have run the [regression tests](https://www.notion.so/Mento-Web-App-Regression-Tests-37bd43a7da8d4e38b65993320a33d557)