Fix: Corrected condition to set quote with swap out amount (#197)
### Description
Corrected the condition to set a quote with the swap-out amount. Before,
the quote was not set for the swap-out amount, and the swap via the
swap-out amount was unavailable (the button was disabled).

### Other changes
None

### Tested
Perform a swap using the 'swap-out' amount.

### Related issues

- Fixes #issue number here

### Checklist before requesting a review

- [x] I have performed a self-review of my own code
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] The PR title follows the
[conventions](https://www.notion.so/Git-Branching-and-Commit-Message-Conventions-18f66f7d06444cfcbac5725ffbc7c04a?pvs=4#9355048863c549ef92fe210a8a1298aa)
- [ ] I have run the [regression
tests](https://www.notion.so/Mento-Web-App-Regression-Tests-37bd43a7da8d4e38b65993320a33d557)

Co-authored-by: Andrew718PLTS <andrew.polovets@mentolabs.xyz>