Refactor: Decompose swap form (#200)
### Description

While trying to debug, I got very annoyed with the unwieldiness of the
SwapForm with lots of components and hooks inlined in the same file.

This PR is about decomposing the SwapForm into smaller parts. It
extracts components and hooks into individual files and aims at making
the DX for this feature better.

### Other changes

- I also upgraded yarn from 3.3.1 to 4.6.0 while I was at it, hence the
`yarn.lock` update, but no dependencies where actually changed or
updated.
- I also spend some time improving the UX of the submit button on the SwapConfirm page and make sure it always displays the right text and the right enabled/disabled status

### Tested

I checked that swaps still work in both directions and with or without
approvals.

### How to review
Everything should work exactly as it does on production right now, no
functionality should have changed.
- [ ] Check out the preview URL
- [ ] Do a few swaps in different directions and with and without
approval TXs
- [ ] Review the code