Fix: Prevented incorrect error handling on getting token balances and corrected cGHS address for Alfajores (#205)
### Description

We couldn't get balances on Alfajores because of the incorrect error
handling on getting token balances - it was trying to get the balance of
the token with the address that is not correct (cGHS, this way we faced
the issue). Now, I replaced it with `try catch` block to prevent
throwing an error when there's an issue with some of the tokens.
Instead, we load all the tokens we can and display the responsive
console error to make it more understandable and prevent problems like
this.

### Other changes
Updated address for the 'cGHS' token.

### Tested
While token/s addresses are correct.
1. Navigate to the preview and connect your wallet where you have tokens
on the Alfajores.
2. Change the network to Alfajores.
3. Wait until the balances are loaded.

While token/s addresses are not correct
1. Check out the branch and change any token address to an incorrect
one.
2. Run locally and connect your wallet where you have tokens on the
Alfajores.
3. Change the network to Alfajores.
4. Wait until the warning notification toast is loaded and the console
error is displayed.

### Related issues

- Fixes #issue number here

### Checklist before requesting a review

- [ ] I have performed a self-review of my own code
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] The PR title follows the
[conventions](https://www.notion.so/Git-Branching-and-Commit-Message-Conventions-18f66f7d06444cfcbac5725ffbc7c04a?pvs=4#9355048863c549ef92fe210a8a1298aa)
- [ ] I have run the [regression
tests](https://www.notion.so/Mento-Web-App-Regression-Tests-37bd43a7da8d4e38b65993320a33d557)

Co-authored-by: chapati <philip.paetz@me.com>